Move emit tests to integration tests directory
This makes it clearer that we actually have some tests! This also hopefully makes it easier to find examples of codespan’s output.
